+++ python/cudf/cudf/core/multiindex.py
@@ -34,13 +34,14 @@ from cudf.core.index import (
     ensure_index,
 )
 from cudf.core.join._join_helpers import _match_join_keys
+from cudf.core.mixins import NotIterable
 from cudf.utils.dtypes import (
     CUDF_STRING_DTYPE,
     SIZE_TYPE_DTYPE,
     is_column_like,
 )
 from cudf.utils.performance_tracking import _performance_tracking
-from cudf.utils.utils import NotIterable, _external_only_api, _is_same_name
+from cudf.utils.utils import _external_only_api, _is_same_name
 
 if TYPE_CHECKING:
     from collections.abc import Generator, Hashable, MutableMapping
